+++ R/orbital.R
@@ -15,6 +15,11 @@
 #'   and `"prob"`. Multiple values are allowed to produce hard and soft
 #'   predictions for classification models. Defaults to `NULL` which defaults to
 #'   `"numeric"` for regression models and `"class"` for classification models.
+#' @param separate_trees A single logical. For tree ensemble models (xgboost,
+#'   lightgbm, catboost, ranger, randomForest), should each tree be output as a
+#'   separate expression? This can improve performance when predicting in
+#'   databases by allowing parallel evaluation of trees. Defaults to `FALSE`.
+#'   See `vignette("separate-trees")` for details.
 #'
 #' @returns An [orbital] object.
 #'
@@ -63,7 +68,13 @@
 #'   orbital()
 #'
 #' @export
-orbital <- function(x, ..., prefix = ".pred", type = NULL) {
+orbital <- function(
+  x,
+  ...,
+  prefix = ".pred",
+  type = NULL,
+  separate_trees = FALSE
+) {
   UseMethod("orbital")
 }